// packages/code-server/src/lens/context.d.ts
export interface LensDB {
    session: {
        findUnique: (args: {
            where: {
                id: string;
            };
        }) => Promise<any>;
        findMany: (args?: {
            where?: any;
            orderBy?: any;
            take?: number;
        }) => Promise<any[]>;
        create: (args: {
            data: any;
        }) => Promise<any>;
        update: (args: {
            where: {
                id: string;
            };
            data: any;
        }) => Promise<any>;
        delete: (args: {
            where: {
                id: string;
            };
        }) => Promise<any>;
        count: () => Promise<number>;
    };
    message: {
        findUnique: (args: {
            where: {
                id: string;
            };
        }) => Promise<any>;
        findMany: (args?: {
            where?: any;
            orderBy?: any;
            take?: number;
        }) => Promise<any[]>;
        create: (args: {
            data: any;
        }) => Promise<any>;
        update: (args: {
            where: {
                id: string;
            };
            data: any;
        }) => Promise<any>;
        delete: (args: {
            where: {
                id: string;
            };
        }) => Promise<any>;
    };
    step: {
        findUnique: (args: {
            where: {
                id: string;
            };
        }) => Promise<any>;
        findMany: (args?: {
            where?: any;
            orderBy?: any;
            take?: number;
        }) => Promise<any[]>;
        create: (args: {
            data: any;
        }) => Promise<any>;
        update: (args: {
            where: {
                id: string;
            };
            data: any;
        }) => Promise<any>;
    };
    part: {
        findUnique: (args: {
            where: {
                id: string;
            };
        }) => Promise<any>;
        findMany: (args?: {
            where?: any;
            orderBy?: any;
            take?: number;
        }) => Promise<any[]>;
        create: (args: {
            data: any;
        }) => Promise<any>;
        update: (args: {
            where: {
                id: string;
            };
            data: any;
        }) => Promise<any>;
    };
    todo: {
        findMany: (args?: {
            where?: any;
            orderBy?: any;
            take?: number;
        }) => Promise<any[]>;
        create: (args: {
            data: any;
        }) => Promise<any>;
        update: (args: {
            where: {
                sessionId: string;
                id: number;
            };
            data: any;
        }) => Promise<any>;
        delete: (args: {
            where: {
                sessionId: string;
                id: number;
            };
        }) => Promise<any>;
    };
}